// src/app/planet/models/planet.model.ts
export interface PlanetModel {
  climate?: string;
  created?: Date;
  diameter?: string | number;
  edited?: Date;
  films?: string[];
  gravity?: string;
  name?: string;
  orbital_period?: string | number;
  population?: string | number;
  residents?: string[];
  rotation_period?: string | number;
  surface_water?: string | number;
  terrain?: string;
  url?: string;
}